parc list and fix some isue in facturelist

+++ A1 GAESTION COMMERCIAL/Controles/ClientRow.vb
@@ -102,6 +102,18 @@
         End Set
     End Property
 
+    Public Sub New()
+
+        ' This call is required by the designer.
+        InitializeComponent()
+
+        ' Add any initialization after the InitializeComponent() call.
+        plP.Width = Form1.cellWidth
+        plQ.Width = Form1.cellWidth
+        plR.Width = Form1.cellWidth
+        plT.Width = Form1.cellWidth
+        plRef.Width = Form1.cellWidth / 2
+    End Sub
     Private Sub PlLeft_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles PlLeft.Click, lbVille.Click,  lbType.Click, lbResponsable.Click, lbref.Click, lbName.Click, lbTel.Click
         isSelected = Not isSelected
         RaiseEvent selected()
@@ -111,7 +123,10 @@
         RaiseEvent EditSelectedItem(Me)
     End Sub
     Private Sub btClear_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btClear.Click
-        RaiseEvent DeleteItem(Me)
+        If MsgBox(MsgDelete & vbNewLine & lbName.Text & " : " & Id, MsgBoxStyle.YesNo, "Suppression") = MsgBoxResult.Yes Then
+            RaiseEvent DeleteItem(Me)
+        End If
+
     End Sub
     Private Sub Button1_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button1.Click
         RaiseEvent GetFactureInfos(Id)
